px4_sitl_packager: add fix for abs paths on Mac
This fixes the case where the absolute path to libmav_msg.dylib had an
absolute path instead of a relative one. It is a very hacky workaround
but it fixes the problem.

A similar issue can be found in:
https://github.com/pandegroup/openmm/issues/1342
but there doesn't seem to be a better fix than this workaround.