Navigator: Fix RTL and Landing (and takeoff gear setting)
Previously we had unnecessary altitude increases in RTL and the gear would only come down after the loiter phase
in RTL and landing. Now RTL does not increase altitude as much and we lower the gear while hovering / loitering
during RTL and landing. This allows to use much lower minimum altitudes and a more intuitive motion in 3D space.
Furthermore the RTL min height is now a cone, requiring the system to not necessarily gain altitude if triggered
close to home. Last the gear is now only raised after takeoff, which is safer for slower takeoffs.