fw image: use git tag as version in firmware file
Write the current git tag ('git describe --tags') in the version field of the
firmware file. This version info can be overwritten by using the '--version'
argument of px_mkfw.py called by the builder. Up to now the version has been
hardcoded to the value set in the json prototype of the firmware file.