fix tap_esc: use uint16_t for motor_out instead of float
This fixes a very dangerous combination between floats and uint16_t bitmasks:
We previously had:
fault tolerant control sets an uint16_t bitmask (reverse bit), which then
was assigned to motor_out (float), which then was converted back to
uint16_t. This really only worked because the float has enough bits to
represent the uint16_t with full precision.

In addition it fixes the bitmasks for RPMMAX & RPMSTOPPED bound checks.