ROMFS: Reset threshold for data link loss to default
This revert part of the change that was done on timeout for RC and data link loss.
It seems that sometimes it take more then 3s to get some packets from the autopilot to the RC.
This cause the user to experience a message of data link loss but the video stream is still present.